requirements.txt
ultralytics==8.0.196
streamlit==1.27.0
opencv-python==4.8.0.76
playsound==1.3.0
pygame==2.5.2
pandas==2.1.1
Pillow==10.0.1
watchdog==3.0.0
pytest==7.4.2
